+++ trunk/TAO/TAO_IDL/be/be_visitor_argument/arglist.cpp
@@ -245,7 +245,7 @@ int be_visitor_args_arglist::visit_string (be_string *node)
 {
   TAO_OutStream *os = this->ctx_->stream (); // get the stream
 
-  if (node->width () == sizeof (char))
+  if (node->width () == (long) sizeof (char))
     {
       switch (this->direction ())
         {
